rename plugin-publisher.ts => publisher.ts + update references
this util is used for both plugin and project publishing so it needs a rename
